@@ -1,11 +1,47 @@
 {
-	"description" : "The CodeIgniter framework",
-	"name" : "codeigniter/framework",
+	"name": "codeigniter4/framework",
+	"type": "project",
+	"description": "The CodeIgniter framework v4",
+	"homepage": "https://codeigniter.com",
 	"license": "MIT",
 	"require": {
-		"php": ">=5.2.4"
+		"php": "^7.3||^8.0",
+		"ext-curl": "*",
+		"ext-intl": "*",
+		"ext-json": "*",
+		"ext-mbstring": "*",
+		"kint-php/kint": "^3.3",
+		"laminas/laminas-escaper": "^2.6",
+		"psr/log": "^1.1"
 	},
 	"require-dev": {
-		"mikey179/vfsStream": "1.1.*"
+		"codeigniter4/codeigniter4-standard": "^1.0",
+		"fakerphp/faker": "^1.9",
+		"mikey179/vfsstream": "^1.6",
+		"phpunit/phpunit": "^9.1",
+		"predis/predis": "^1.1",
+		"squizlabs/php_codesniffer": "^3.3"
+	},
+	"suggest": {
+		"ext-fileinfo": "Improves mime type detection for files"
+	},
+	"autoload": {
+		"psr-4": {
+			"CodeIgniter\\": "system/"
+		},
+		"exclude-from-classmap": [
+			"**/Database/Migrations/**"
+		]
+	},
+	"scripts": {
+		"post-update-cmd": [
+			"CodeIgniter\\ComposerScripts::postUpdate"
+		],
+		"test": "phpunit"
+	},
+	"support": {
+		"forum": "http://forum.codeigniter.com/",
+		"source": "https://github.com/codeigniter4/CodeIgniter4",
+		"slack": "https://codeigniterchat.slack.com"
 	}
 }
