@@ -1,185 +1,279 @@
 <?php
+
 /**
- * CodeIgniter
+ * This file is part of the CodeIgniter 4 framework.
  *
- * An open source application development framework for PHP
+ * (c) CodeIgniter Foundation <admin@codeigniter.com>
  *
- * This content is released under the MIT License (MIT)
- *
- * Copyright (c) 2014 - 2015, British Columbia Institute of Technology
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- * @package	CodeIgniter
- * @author	EllisLab Dev Team
- * @copyright	Copyright (c) 2008 - 2014, EllisLab, Inc. (http://ellislab.com/)
- * @copyright	Copyright (c) 2014 - 2015, British Columbia Institute of Technology (http://bcit.ca/)
- * @license	http://opensource.org/licenses/MIT	MIT License
- * @link	http://codeigniter.com
- * @since	Version 1.0.0
- * @filesource
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
-defined('BASEPATH') OR exit('No direct script access allowed');
+
+use CodeIgniter\HTTP\IncomingRequest;
+use CodeIgniter\HTTP\URI;
+use CodeIgniter\Router\Exceptions\RouterException;
+use Config\App;
+use Config\Services;
 
 /**
  * CodeIgniter URL Helpers
- *
- * @package		CodeIgniter
- * @subpackage	Helpers
- * @category	Helpers
- * @author		EllisLab Dev Team
- * @link		http://codeigniter.com/user_guide/helpers/url_helper.html
  */
 
-// ------------------------------------------------------------------------
-
-if ( ! function_exists('site_url'))
+if (! function_exists('_get_uri'))
 {
 	/**
-	 * Site URL
+	 * Used by the other URL functions to build a
+	 * framework-specific URI based on the App config.
 	 *
-	 * Create a local URL based on your basepath. Segments can be passed via the
-	 * first parameter either as a string or an array.
+	 * @internal Outside of the framework this should not be used directly.
 	 *
-	 * @param	string	$uri
-	 * @param	string	$protocol
-	 * @return	string
+	 * @param string   $relativePath May include queries or fragments
+	 * @param App|null $config
+	 *
+	 * @return URI
+	 *
+	 * @throws InvalidArgumentException For invalid paths or config
 	 */
-	function site_url($uri = '', $protocol = NULL)
+	function _get_uri(string $relativePath = '', App $config = null): URI
 	{
-		return get_instance()->config->site_url($uri, $protocol);
+		$config = $config ?? config('App');
+
+		if ($config->baseURL === '')
+		{
+			throw new InvalidArgumentException('_get_uri() requires a valid baseURL.');
+		}
+
+		// If a full URI was passed then convert it
+		if (is_int(strpos($relativePath, '://')))
+		{
+			$full         = new URI($relativePath);
+			$relativePath = URI::createURIString(null, null, $full->getPath(), $full->getQuery(), $full->getFragment());
+		}
+
+		$relativePath = URI::removeDotSegments($relativePath);
+
+		// Build the full URL based on $config and $relativePath
+		$url = rtrim($config->baseURL, '/ ') . '/';
+
+		// Check for an index page
+		if ($config->indexPage !== '')
+		{
+			$url .= $config->indexPage;
+
+			// Check if we need a separator
+			if ($relativePath !== '' && $relativePath[0] !== '/' && $relativePath[0] !== '?')
+			{
+				$url .= '/';
+			}
+		}
+
+		$url .= $relativePath;
+
+		$uri = new URI($url);
+
+		// Check if the baseURL scheme needs to be coerced into its secure version
+		if ($config->forceGlobalSecureRequests && $uri->getScheme() === 'http')
+		{
+			$uri->setScheme('https');
+		}
+
+		return $uri;
 	}
 }
 
-// ------------------------------------------------------------------------
+//--------------------------------------------------------------------
 
-if ( ! function_exists('base_url'))
+if (! function_exists('site_url'))
 {
 	/**
-	 * Base URL
+	 * Returns a site URL as defined by the App config.
 	 *
-	 * Create a local URL based on your basepath.
-	 * Segments can be passed in as a string or an array, same as site_url
-	 * or a URL to a file can be passed in, e.g. to an image file.
+	 * @param mixed       $relativePath URI string or array of URI segments
+	 * @param string|null $scheme
+	 * @param App|null    $config       Alternate configuration to use
 	 *
-	 * @param	string	$uri
-	 * @param	string	$protocol
-	 * @return	string
+	 * @return string
 	 */
-	function base_url($uri = '', $protocol = NULL)
+	function site_url($relativePath = '', string $scheme = null, App $config = null): string
 	{
-		return get_instance()->config->base_url($uri, $protocol);
+		// Convert array of segments to a string
+		if (is_array($relativePath))
+		{
+			$relativePath = implode('/', $relativePath);
+		}
+
+		$uri = _get_uri($relativePath, $config);
+
+		return URI::createURIString($scheme ?? $uri->getScheme(), $uri->getAuthority(), $uri->getPath(), $uri->getQuery(), $uri->getFragment());
 	}
 }
 
-// ------------------------------------------------------------------------
+//--------------------------------------------------------------------
 
-if ( ! function_exists('current_url'))
+if (! function_exists('base_url'))
 {
 	/**
-	 * Current URL
+	 * Returns the base URL as defined by the App config.
+	 * Base URLs are trimmed site URLs without the index page.
 	 *
-	 * Returns the full URL (including segments) of the page where this
-	 * function is placed
-	 *
-	 * @return	string
+	 * @param  mixed  $relativePath URI string or array of URI segments
+	 * @param  string $scheme
+	 * @return string
 	 */
-	function current_url()
+	function base_url($relativePath = '', string $scheme = null): string
 	{
-		$CI =& get_instance();
-		return $CI->config->site_url($CI->uri->uri_string());
+		$config            = clone config('App');
+		$config->indexPage = '';
+
+		return rtrim(site_url($relativePath, $scheme, $config), '/');
 	}
 }
 
-// ------------------------------------------------------------------------
+//--------------------------------------------------------------------
 
-if ( ! function_exists('uri_string'))
+if (! function_exists('current_url'))
+{
+	/**
+	 * Returns the current full URL based on the IncomingRequest.
+	 * String returns ignore query and fragment parts.
+	 *
+	 * @param boolean              $returnObject True to return an object instead of a string
+	 * @param IncomingRequest|null $request      A request to use when retrieving the path
+	 *
+	 * @return string|URI
+	 */
+	function current_url(bool $returnObject = false, IncomingRequest $request = null)
+	{
+		$request = $request ?? Services::request();
+		$path    = $request->getPath();
+
+		// Append queries and fragments
+		if ($query = $request->getUri()->getQuery())
+		{
+			$path .= '?' . $query;
+		}
+		if ($fragment = $request->getUri()->getFragment())
+		{
+			$path .= '#' . $fragment;
+		}
+
+		$uri = _get_uri($path);
+
+		return $returnObject ? $uri : URI::createURIString($uri->getScheme(), $uri->getAuthority(), $uri->getPath());
+	}
+}
+
+//--------------------------------------------------------------------
+
+if (! function_exists('previous_url'))
+{
+	/**
+	 * Returns the previous URL the current visitor was on. For security reasons
+	 * we first check in a saved session variable, if it exists, and use that.
+	 * If that's not available, however, we'll use a sanitized url from $_SERVER['HTTP_REFERER']
+	 * which can be set by the user so is untrusted and not set by certain browsers/servers.
+	 *
+	 * @param boolean $returnObject
+	 *
+	 * @return URI|mixed|string
+	 */
+	function previous_url(bool $returnObject = false)
+	{
+		// Grab from the session first, if we have it,
+		// since it's more reliable and safer.
+		// Otherwise, grab a sanitized version from $_SERVER.
+		$referer = $_SESSION['_ci_previous_url'] ?? Services::request()->getServer('HTTP_REFERER', FILTER_SANITIZE_URL);
+
+		$referer = $referer ?? site_url('/');
+
+		return $returnObject ? new URI($referer) : $referer;
+	}
+}
+
+//--------------------------------------------------------------------
+
+if (! function_exists('uri_string'))
 {
 	/**
 	 * URL String
 	 *
-	 * Returns the URI segments.
+	 * Returns the path part of the current URL
 	 *
-	 * @return	string
+	 * @param boolean $relative Whether the resulting path should be relative to baseURL
+	 *
+	 * @return string
 	 */
-	function uri_string()
+	function uri_string(bool $relative = false): string
 	{
-		return get_instance()->uri->uri_string();
+		return $relative
+			? ltrim(Services::request()->getPath(), '/')
+			: Services::request()->getUri()->getPath();
 	}
 }
 
-// ------------------------------------------------------------------------
+//--------------------------------------------------------------------
 
-if ( ! function_exists('index_page'))
+if (! function_exists('index_page'))
 {
 	/**
 	 * Index page
 	 *
 	 * Returns the "index_page" from your config file
 	 *
-	 * @return	string
+	 * @param  App|null $altConfig Alternate configuration to use
+	 * @return string
 	 */
-	function index_page()
+	function index_page(App $altConfig = null): string
 	{
-		return get_instance()->config->item('index_page');
+		// use alternate config if provided, else default one
+		$config = $altConfig ?? config(App::class);
+
+		return $config->indexPage;
 	}
 }
 
 // ------------------------------------------------------------------------
 
-if ( ! function_exists('anchor'))
+if (! function_exists('anchor'))
 {
 	/**
 	 * Anchor Link
 	 *
 	 * Creates an anchor based on the local URL.
 	 *
-	 * @param	string	the URL
-	 * @param	string	the link title
-	 * @param	mixed	any attributes
-	 * @return	string
+	 * @param mixed    $uri        URI string or array of URI segments
+	 * @param string   $title      The link title
+	 * @param mixed    $attributes Any attributes
+	 * @param App|null $altConfig  Alternate configuration to use
+	 *
+	 * @return string
 	 */
-	function anchor($uri = '', $title = '', $attributes = '')
+	function anchor($uri = '', string $title = '', $attributes = '', App $altConfig = null): string
 	{
-		$title = (string) $title;
+		// use alternate config if provided, else default one
+		$config = $altConfig ?? config(App::class);
 
-		$site_url = is_array($uri)
-			? site_url($uri)
-			: preg_match('#^(\w+:)?//#i', $uri) ? $uri : site_url($uri);
+		$siteUrl = is_array($uri) ? site_url($uri, null, $config) : (preg_match('#^(\w+:)?//#i', $uri) ? $uri : site_url($uri, null, $config));
+		// eliminate trailing slash
+		$siteUrl = rtrim($siteUrl, '/');
 
 		if ($title === '')
 		{
-			$title = $site_url;
+			$title = $siteUrl;
 		}
 
 		if ($attributes !== '')
 		{
-			$attributes = _stringify_attributes($attributes);
+			$attributes = stringify_attributes($attributes);
 		}
 
-		return '<a href="'.$site_url.'"'.$attributes.'>'.$title.'</a>';
+		return '<a href="' . $siteUrl . '"' . $attributes . '>' . $title . '</a>';
 	}
 }
 
 // ------------------------------------------------------------------------
 
-if ( ! function_exists('anchor_popup'))
+if (! function_exists('anchor_popup'))
 {
 	/**
 	 * Anchor Link - Pop-up version
@@ -187,110 +281,113 @@ if ( ! function_exists('anchor_popup'))
 	 * Creates an anchor based on the local URL. The link
 	 * opens a new window based on the attributes specified.
 	 *
-	 * @param	string	the URL
-	 * @param	string	the link title
-	 * @param	mixed	any attributes
-	 * @return	string
+	 * @param string   $uri        the URL
+	 * @param string   $title      the link title
+	 * @param mixed    $attributes any attributes
+	 * @param App|null $altConfig  Alternate configuration to use
+	 *
+	 * @return string
 	 */
-	function anchor_popup($uri = '', $title = '', $attributes = FALSE)
+	function anchor_popup($uri = '', string $title = '', $attributes = false, App $altConfig = null): string
 	{
-		$title = (string) $title;
-		$site_url = preg_match('#^(\w+:)?//#i', $uri) ? $uri : site_url($uri);
+		// use alternate config if provided, else default one
+		$config = $altConfig ?? config(App::class);
+
+		$siteUrl = preg_match('#^(\w+:)?//#i', $uri) ? $uri : site_url($uri, null, $config);
+		$siteUrl = rtrim($siteUrl, '/');
 
 		if ($title === '')
 		{
-			$title = $site_url;
+			$title = $siteUrl;
 		}
 
-		if ($attributes === FALSE)
+		if ($attributes === false)
 		{
-			return '<a href="'.$site_url.'" onclick="window.open(\''.$site_url."', '_blank'); return false;\">".$title.'</a>';
+			return '<a href="' . $siteUrl . '" onclick="window.open(\'' . $siteUrl . "', '_blank'); return false;\">" . $title . '</a>';
 		}
 
-		if ( ! is_array($attributes))
+		if (! is_array($attributes))
 		{
-			$attributes = array($attributes);
+			$attributes = [$attributes];
 
 			// Ref: http://www.w3schools.com/jsref/met_win_open.asp
-			$window_name = '_blank';
+			$windowName = '_blank';
 		}
-		elseif ( ! empty($attributes['window_name']))
+		elseif (! empty($attributes['window_name']))
 		{
-			$window_name = $attributes['window_name'];
+			$windowName = $attributes['window_name'];
 			unset($attributes['window_name']);
 		}
 		else
 		{
-			$window_name = '_blank';
+			$windowName = '_blank';
 		}
 
-		foreach (array('width' => '800', 'height' => '600', 'scrollbars' => 'yes', 'menubar' => 'no', 'status' => 'yes', 'resizable' => 'yes', 'screenx' => '0', 'screeny' => '0') as $key => $val)
+		foreach (['width' => '800', 'height' => '600', 'scrollbars' => 'yes', 'menubar' => 'no', 'status' => 'yes', 'resizable' => 'yes', 'screenx' => '0', 'screeny' => '0'] as $key => $val)
 		{
-			$atts[$key] = isset($attributes[$key]) ? $attributes[$key] : $val;
+			$atts[$key] = $attributes[$key] ?? $val;
 			unset($attributes[$key]);
 		}
 
-		$attributes = _stringify_attributes($attributes);
+		$attributes = stringify_attributes($attributes);
 
-		return '<a href="'.$site_url
-			.'" onclick="window.open(\''.$site_url."', '".$window_name."', '"._stringify_attributes($atts, TRUE)."'); return false;\""
-			.$attributes.'>'.$title.'</a>';
+		return '<a href="' . $siteUrl
+				. '" onclick="window.open(\'' . $siteUrl . "', '" . $windowName . "', '" . stringify_attributes($atts, true) . "'); return false;\""
+				. $attributes . '>' . $title . '</a>';
 	}
 }
 
 // ------------------------------------------------------------------------
 
-if ( ! function_exists('mailto'))
+if (! function_exists('mailto'))
 {
 	/**
 	 * Mailto Link
 	 *
-	 * @param	string	the email address
-	 * @param	string	the link title
-	 * @param	mixed	any attributes
-	 * @return	string
+	 * @param string $email      the email address
+	 * @param string $title      the link title
+	 * @param mixed  $attributes any attributes
+	 *
+	 * @return string
 	 */
-	function mailto($email, $title = '', $attributes = '')
+	function mailto(string $email, string $title = '', $attributes = ''): string
 	{
-		$title = (string) $title;
-
-		if ($title === '')
+		if (trim($title) === '')
 		{
 			$title = $email;
 		}
 
-		return '<a href="mailto:'.$email.'"'._stringify_attributes($attributes).'>'.$title.'</a>';
+		return '<a href="mailto:' . $email . '"' . stringify_attributes($attributes) . '>' . $title . '</a>';
 	}
 }
 
 // ------------------------------------------------------------------------
 
-if ( ! function_exists('safe_mailto'))
+if (! function_exists('safe_mailto'))
 {
 	/**
 	 * Encoded Mailto Link
 	 *
 	 * Create a spam-protected mailto link written in Javascript
 	 *
-	 * @param	string	the email address
-	 * @param	string	the link title
-	 * @param	mixed	any attributes
-	 * @return	string
+	 * @param string $email      the email address
+	 * @param string $title      the link title
+	 * @param mixed  $attributes any attributes
+	 *
+	 * @return string
 	 */
-	function safe_mailto($email, $title = '', $attributes = '')
+	function safe_mailto(string $email, string $title = '', $attributes = ''): string
 	{
-		$title = (string) $title;
-
-		if ($title === '')
+		if (trim($title) === '')
 		{
 			$title = $email;
 		}
 
 		$x = str_split('<a href="mailto:', 1);
 
-		for ($i = 0, $l = strlen($email); $i < $l; $i++)
+		for ($i = 0, $l = strlen($email); $i < $l; $i ++)
 		{
-			$x[] = '|'.ord($email[$i]);
+			$x[] = '|' . ord($email[$i]);
 		}
 
 		$x[] = '"';
@@ -301,81 +398,79 @@ if ( ! function_exists('safe_mailto'))
 			{
 				foreach ($attributes as $key => $val)
 				{
-					$x[] = ' '.$key.'="';
-					for ($i = 0, $l = strlen($val); $i < $l; $i++)
+					$x[] = ' ' . $key . '="';
+					for ($i = 0, $l = strlen($val); $i < $l; $i ++)
 					{
-						$x[] = '|'.ord($val[$i]);
+						$x[] = '|' . ord($val[$i]);
 					}
 					$x[] = '"';
 				}
 			}
 			else
 			{
-				for ($i = 0, $l = strlen($attributes); $i < $l; $i++)
+				for ($i = 0, $l = mb_strlen($attributes); $i < $l; $i ++)
 				{
-					$x[] = $attributes[$i];
+					$x[] = mb_substr($attributes, $i, 1);
 				}
 			}
 		}
 
 		$x[] = '>';
 
-		$temp = array();
-		for ($i = 0, $l = strlen($title); $i < $l; $i++)
+		$temp = [];
+		for ($i = 0, $l = strlen($title); $i < $l; $i ++)
 		{
 			$ordinal = ord($title[$i]);
 
 			if ($ordinal < 128)
 			{
-				$x[] = '|'.$ordinal;
+				$x[] = '|' . $ordinal;
 			}
 			else
 			{
-				if (count($temp) === 0)
+				if (empty($temp))
 				{
 					$count = ($ordinal < 224) ? 2 : 3;
 				}
 
 				$temp[] = $ordinal;
-				if (count($temp) === $count)
+				if (count($temp) === $count) // @phpstan-ignore-line
 				{
-					$number = ($count === 3)
-							? (($temp[0] % 16) * 4096) + (($temp[1] % 64) * 64) + ($temp[2] % 64)
-							: (($temp[0] % 32) * 64) + ($temp[1] % 64);
-					$x[] = '|'.$number;
-					$count = 1;
-					$temp = array();
+					$number = ($count === 3) ? (($temp[0] % 16) * 4096) + (($temp[1] % 64) * 64) + ($temp[2] % 64) : (($temp[0] % 32) * 64) + ($temp[1] % 64);
+					$x[]    = '|' . $number;
+					$count  = 1;
+					$temp   = [];
 				}
 			}
 		}
 
-		$x[] = '<'; $x[] = '/'; $x[] = 'a'; $x[] = '>';
+		$x[] = '<';
+		$x[] = '/';
+		$x[] = 'a';
+		$x[] = '>';
 
 		$x = array_reverse($x);
 
-		$output = "<script type=\"text/javascript\">\n"
-			."\t//<![CDATA[\n"
-			."\tvar l=new Array();\n";
+		// improve obfuscation by eliminating newlines & whitespace
+		$output = '<script type="text/javascript">'
+				. 'var l=new Array();';
 
-		for ($i = 0, $c = count($x); $i < $c; $i++)
+		foreach ($x as $i => $value)
 		{
-			$output .= "\tl[".$i."] = '".$x[$i]."';\n";
+			$output .= 'l[' . $i . "] = '" . $value . "';";
 		}
 
-		$output .= "\n\tfor (var i = l.length-1; i >= 0; i=i-1) {\n"
-			."\t\tif (l[i].substring(0, 1) === '|') document.write(\"&#\"+unescape(l[i].substring(1))+\";\");\n"
-			."\t\telse document.write(unescape(l[i]));\n"
-			."\t}\n"
-			."\t//]]>\n"
-			.'</script>';
-
-		return $output;
+		return $output . ('for (var i = l.length-1; i >= 0; i=i-1) {'
+				. "if (l[i].substring(0, 1) === '|') document.write(\"&#\"+unescape(l[i].substring(1))+\";\");"
+				. 'else document.write(unescape(l[i]));'
+				. '}'
+				. '</script>');
 	}
 }
 
 // ------------------------------------------------------------------------
 
-if ( ! function_exists('auto_link'))
+if (! function_exists('auto_link'))
 {
 	/**
 	 * Auto-linker
@@ -385,12 +480,13 @@ if ( ! function_exists('auto_link'))
 	 * URLs or emails that end in a period. We'll strip these
 	 * off and add them after the link.
 	 *
-	 * @param	string	the string
-	 * @param	string	the type: email, url, or both
-	 * @param	bool	whether to create pop-up links
-	 * @return	string
+	 * @param string  $str   the string
+	 * @param string  $type  the type: email, url, or both
+	 * @param boolean $popup whether to create pop-up links
+	 *
+	 * @return string
 	 */
-	function auto_link($str, $type = 'both', $popup = FALSE)
+	function auto_link(string $str, string $type = 'both', bool $popup = false): string
 	{
 		// Find and replace any URLs.
 		if ($type !== 'email' && preg_match_all('#(\w*://|www\.)[^\s()<>;]+\w#i', $str, $matches, PREG_OFFSET_CAPTURE | PREG_SET_ORDER))
@@ -408,7 +504,7 @@ if ( ! function_exists('auto_link'))
 				//
 				// With PREG_OFFSET_CAPTURE, both of the above is an array,
 				// where the actual value is held in [0] and its offset at the [1] index.
-				$a = '<a href="'.(strpos($match[1][0], '/') ? '' : 'http://').$match[0][0].'"'.$target.'>'.$match[0][0].'</a>';
+				$a   = '<a href="' . (strpos($match[1][0], '/') ? '' : 'http://') . $match[0][0] . '"' . $target . '>' . $match[0][0] . '</a>';
 				$str = substr_replace($str, $a, $match[0][1], strlen($match[0][0]));
 			}
 		}
@@ -418,7 +514,7 @@ if ( ! function_exists('auto_link'))
 		{
 			foreach (array_reverse($matches[0]) as $match)
 			{
-				if (filter_var($match[0], FILTER_VALIDATE_EMAIL) !== FALSE)
+				if (filter_var($match[0], FILTER_VALIDATE_EMAIL) !== false)
 				{
 					$str = substr_replace($str, safe_mailto($match[0]), $match[1], strlen($match[0]));
 				}
@@ -431,28 +527,34 @@ if ( ! function_exists('auto_link'))
 
 // ------------------------------------------------------------------------
 
-if ( ! function_exists('prep_url'))
+if (! function_exists('prep_url'))
 {
 	/**
-	 * Prep URL
+	 * Prep URL - Simply adds the http:// or https:// part if no scheme is included.
 	 *
-	 * Simply adds the http:// part if no scheme is included
+	 * Formerly used URI, but that does not play nicely with URIs missing
+	 * the scheme.
 	 *
-	 * @param	string	the URL
-	 * @return	string
+	 * @param  string  $str    the URL
+	 * @param  boolean $secure set true if you want to force https://
+	 * @return string
 	 */
-	function prep_url($str = '')
+	function prep_url(string $str = '', bool $secure = false): string
 	{
-		if ($str === 'http://' OR $str === '')
+		if (in_array($str, ['http://', 'https://', '//', ''], true))
 		{
 			return '';
 		}
 
-		$url = parse_url($str);
-
-		if ( ! $url OR ! isset($url['scheme']))
+		if (parse_url($str, PHP_URL_SCHEME) === null)
 		{
-			return 'http://'.$str;
+			$str = 'http://' . ltrim($str, '/');
+		}
+
+		// force replace http:// with https://
+		if ($secure)
+		{
+			$str = preg_replace('/^(?:http):/i', 'https:', $str);
 		}
 
 		return $str;
@@ -461,7 +563,7 @@ if ( ! function_exists('prep_url'))
 
 // ------------------------------------------------------------------------
 
-if ( ! function_exists('url_title'))
+if (! function_exists('url_title'))
 {
 	/**
 	 * Create URL Title
@@ -470,42 +572,31 @@ if ( ! function_exists('url_title'))
 	 * human-friendly URL string with a "separator" string
 	 * as the word separator.
 	 *
-	 * @todo	Remove old 'dash' and 'underscore' usage in 3.1+.
-	 * @param	string	$str		Input string
-	 * @param	string	$separator	Word separator
-	 *			(usually '-' or '_')
-	 * @param	bool	$lowercase	Wether to transform the output string to lowercase
-	 * @return	string
+	 * @param  string  $str       Input string
+	 * @param  string  $separator Word separator (usually '-' or '_')
+	 * @param  boolean $lowercase Whether to transform the output string to lowercase
+	 * @return string
 	 */
-	function url_title($str, $separator = '-', $lowercase = FALSE)
+	function url_title(string $str, string $separator = '-', bool $lowercase = false): string
 	{
-		if ($separator === 'dash')
-		{
-			$separator = '-';
-		}
-		elseif ($separator === 'underscore')
-		{
-			$separator = '_';
-		}
+		$qSeparator = preg_quote($separator, '#');
 
-		$q_separator = preg_quote($separator, '#');
-
-		$trans = array(
-			'&.+?;'			=> '',
-			'[^a-z0-9 _-]'		=> '',
-			'\s+'			=> $separator,
-			'('.$q_separator.')+'	=> $separator
-		);
+		$trans = [
+			'&.+?;'                  => '',
+			'[^\w\d\pL\pM _-]'       => '',
+			'\s+'                    => $separator,
+			'(' . $qSeparator . ')+' => $separator,
+		];
 
 		$str = strip_tags($str);
 		foreach ($trans as $key => $val)
 		{
-			$str = preg_replace('#'.$key.'#i', $val, $str);
+			$str = preg_replace('#' . $key . '#iu', $val, $str);
 		}
 
-		if ($lowercase === TRUE)
+		if ($lowercase === true)
 		{
-			$str = strtolower($str);
+			$str = mb_strtolower($str);
 		}
 
 		return trim(trim($str, $separator));
@@ -514,56 +605,81 @@ if ( ! function_exists('url_title'))
 
 // ------------------------------------------------------------------------
 
-if ( ! function_exists('redirect'))
+if (! function_exists('mb_url_title'))
 {
 	/**
-	 * Header Redirect
+	 * Create URL Title that takes into account accented characters
 	 *
-	 * Header redirect in two flavors
-	 * For very fine grained control over headers, you could use the Output
-	 * Library's set_header() function.
+	 * Takes a "title" string as input and creates a
+	 * human-friendly URL string with a "separator" string
+	 * as the word separator.
 	 *
-	 * @param	string	$uri	URL
-	 * @param	string	$method	Redirect method
-	 *			'auto', 'location' or 'refresh'
-	 * @param	int	$code	HTTP Response status code
-	 * @return	void
+	 * @param  string  $str       Input string
+	 * @param  string  $separator Word separator (usually '-' or '_')
+	 * @param  boolean $lowercase Whether to transform the output string to lowercase
+	 * @return string
 	 */
-	function redirect($uri = '', $method = 'auto', $code = NULL)
+	function mb_url_title(string $str, string $separator = '-', bool $lowercase = false): string
 	{
-		if ( ! preg_match('#^(\w+:)?//#i', $uri))
+		helper('text');
+
+		return url_title(convert_accented_characters($str), $separator, $lowercase);
+	}
+}
+
+//--------------------------------------------------------------------
+
+if (! function_exists('url_to'))
+{
+	/**
+	 * Get the full, absolute URL to a controller method
+	 * (with additional arguments)
+	 *
+	 * @param string $controller
+	 * @param mixed  ...$args
+	 *
+	 * @throws RouterException
+	 *
+	 * @return string
+	 */
+	function url_to(string $controller, ...$args): string
+	{
+		if (! $route = route_to($controller, ...$args))
 		{
-			$uri = site_url($uri);
+			$explode = explode('::', $controller);
+
+			if (isset($explode[1]))
+			{
+				throw RouterException::forControllerNotFound($explode[0], $explode[1]);
+			}
+
+			throw RouterException::forInvalidRoute($controller);
 		}
 
-		// IIS environment likely? Use 'refresh' for better compatibility
-		if ($method === 'auto' && isset($_SERVER['SERVER_SOFTWARE']) && strpos($_SERVER['SERVER_SOFTWARE'], 'Microsoft-IIS') !== FALSE)
-		{
-			$method = 'refresh';
-		}
-		elseif ($method !== 'refresh' && (empty($code) OR ! is_numeric($code)))
-		{
-			if (isset($_SERVER['SERVER_PROTOCOL'], $_SERVER['REQUEST_METHOD']) && $_SERVER['SERVER_PROTOCOL'] === 'HTTP/1.1')
-			{
-				$code = ($_SERVER['REQUEST_METHOD'] !== 'GET')
-					? 303	// reference: http://en.wikipedia.org/wiki/Post/Redirect/Get
-					: 307;
-			}
-			else
-			{
-				$code = 302;
-			}
-		}
+		return site_url($route);
+	}
+}
 
-		switch ($method)
-		{
-			case 'refresh':
-				header('Refresh:0;url='.$uri);
-				break;
-			default:
-				header('Location: '.$uri, TRUE, $code);
-				break;
-		}
-		exit;
+if (! function_exists('url_is'))
+{
+	/**
+	 * Determines if current url path contains
+	 * the given path. It may contain a wildcard (*)
+	 * which will allow any valid character.
+	 *
+	 * Example:
+	 *   if (url_is('admin*)) ...
+	 *
+	 * @param string $path
+	 *
+	 * @return boolean
+	 */
+	function url_is(string $path): bool
+	{
+		// Setup our regex to allow wildcards
+		$path        = '/' . trim(str_replace('*', '(\S)*', $path), '/ ');
+		$currentPath = '/' . trim(uri_string(true), '/ ');
+
+		return (bool) preg_match("|^{$path}$|", $currentPath, $matches);
 	}
 }
