@@ -1,55 +1,22 @@
 <?php
-/**
- * CodeIgniter
- *
- * An open source application development framework for PHP
- *
- * This content is released under the MIT License (MIT)
- *
- * Copyright (c) 2014 - 2015, British Columbia Institute of Technology
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- * @package	CodeIgniter
- * @author	EllisLab Dev Team
- * @copyright	Copyright (c) 2008 - 2014, EllisLab, Inc. (http://ellislab.com/)
- * @copyright	Copyright (c) 2014 - 2015, British Columbia Institute of Technology (http://bcit.ca/)
- * @license	http://opensource.org/licenses/MIT	MIT License
- * @link	http://codeigniter.com
- * @since	Version 1.0.0
- * @filesource
- */
-defined('BASEPATH') OR exit('No direct script access allowed');
 
 /**
- * CodeIgniter Cookie Helpers
+ * This file is part of the CodeIgniter 4 framework.
  *
- * @package		CodeIgniter
- * @subpackage	Helpers
- * @category	Helpers
- * @author		EllisLab Dev Team
- * @link		http://codeigniter.com/user_guide/helpers/cookie_helper.html
+ * (c) CodeIgniter Foundation <admin@codeigniter.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
 
-// ------------------------------------------------------------------------
+use Config\App;
+use Config\Services;
 
-if ( ! function_exists('set_cookie'))
+//=============================================================================
+// CodeIgniter Cookie Helpers
+//=============================================================================
+
+if (! function_exists('set_cookie'))
 {
 	/**
 	 * Set cookie
@@ -57,57 +24,90 @@ if ( ! function_exists('set_cookie'))
 	 * Accepts seven parameters, or you can submit an associative
 	 * array in the first parameter containing all the values.
 	 *
-	 * @param	mixed
-	 * @param	string	the value of the cookie
-	 * @param	string	the number of seconds until expiration
-	 * @param	string	the cookie domain.  Usually:  .yourdomain.com
-	 * @param	string	the cookie path
-	 * @param	string	the cookie prefix
-	 * @param	bool	true makes the cookie secure
-	 * @param	bool	true makes the cookie accessible via http(s) only (no javascript)
-	 * @return	void
+	 * @param string|array $name     Cookie name or array containing binds
+	 * @param string       $value    The value of the cookie
+	 * @param string       $expire   The number of seconds until expiration
+	 * @param string       $domain   For site-wide cookie. Usually: .yourdomain.com
+	 * @param string       $path     The cookie path
+	 * @param string       $prefix   The cookie prefix
+	 * @param boolean      $secure   True makes the cookie secure
+	 * @param boolean      $httpOnly True makes the cookie accessible via http(s) only (no javascript)
+	 * @param string|null  $sameSite The cookie SameSite value
+	 *
+	 * @see \CodeIgniter\HTTP\Response::setCookie()
 	 */
-	function set_cookie($name, $value = '', $expire = '', $domain = '', $path = '/', $prefix = '', $secure = FALSE, $httponly = FALSE)
+	function set_cookie(
+		$name,
+		string $value = '',
+		string $expire = '',
+		string $domain = '',
+		string $path = '/',
+		string $prefix = '',
+		bool $secure = false,
+		bool $httpOnly = false,
+		string $sameSite = null
+	)
 	{
-		// Set the config file options
-		get_instance()->input->set_cookie($name, $value, $expire, $domain, $path, $prefix, $secure, $httponly);
+		$response = Services::response();
+		$response->setCookie($name, $value, $expire, $domain, $path, $prefix, $secure, $httpOnly, $sameSite);
 	}
 }
 
-// --------------------------------------------------------------------
-
-if ( ! function_exists('get_cookie'))
+if (! function_exists('get_cookie'))
 {
 	/**
-	 * Fetch an item from the COOKIE array
+	 * Fetch an item from the $_COOKIE array
 	 *
-	 * @param	string
-	 * @param	bool
-	 * @return	mixed
+	 * @param string  $index
+	 * @param boolean $xssClean
+	 *
+	 * @return mixed
+	 *
+	 * @see \CodeIgniter\HTTP\IncomingRequest::getCookie()
 	 */
-	function get_cookie($index, $xss_clean = NULL)
+	function get_cookie($index, bool $xssClean = false)
 	{
-		is_bool($xss_clean) OR $xss_clean = (config_item('global_xss_filtering') === TRUE);
-		$prefix = isset($_COOKIE[$index]) ? '' : config_item('cookie_prefix');
-		return get_instance()->input->cookie($prefix.$index, $xss_clean);
+		$prefix  = isset($_COOKIE[$index]) ? '' : config(App::class)->cookiePrefix;
+		$request = Services::request();
+		$filter  = $xssClean ? FILTER_SANITIZE_STRING : FILTER_DEFAULT;
+
+		return $request->getCookie($prefix . $index, $filter);
 	}
 }
 
-// --------------------------------------------------------------------
-
-if ( ! function_exists('delete_cookie'))
+if (! function_exists('delete_cookie'))
 {
 	/**
-	 * Delete a COOKIE
+	 * Delete a cookie
 	 *
-	 * @param	mixed
-	 * @param	string	the cookie domain. Usually: .yourdomain.com
-	 * @param	string	the cookie path
-	 * @param	string	the cookie prefix
-	 * @return	void
+	 * @param mixed  $name
+	 * @param string $domain the cookie domain. Usually: .yourdomain.com
+	 * @param string $path   the cookie path
+	 * @param string $prefix the cookie prefix
+	 *
+	 * @return void
+	 *
+	 * @see \CodeIgniter\HTTP\Response::deleteCookie()
 	 */
-	function delete_cookie($name, $domain = '', $path = '/', $prefix = '')
+	function delete_cookie($name, string $domain = '', string $path = '/', string $prefix = '')
 	{
-		set_cookie($name, '', '', $domain, $path, $prefix);
+		Services::response()->deleteCookie($name, $domain, $path, $prefix);
+	}
+}
+
+if (! function_exists('has_cookie'))
+{
+	/**
+	 * Checks if a cookie exists by name.
+	 *
+	 * @param string      $name
+	 * @param string|null $value
+	 * @param string      $prefix
+	 *
+	 * @return boolean
+	 */
+	function has_cookie(string $name, string $value = null, string $prefix = ''): bool
+	{
+		return Services::response()->hasCookie($name, $value, $prefix);
 	}
 }
