@@ -1,137 +1,71 @@
 <?php
+
 /**
- * CodeIgniter
+ * This file is part of the CodeIgniter 4 framework.
  *
- * An open source application development framework for PHP
+ * (c) CodeIgniter Foundation <admin@codeigniter.com>
  *
- * This content is released under the MIT License (MIT)
- *
- * Copyright (c) 2014 - 2015, British Columbia Institute of Technology
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- * @package	CodeIgniter
- * @author	EllisLab Dev Team
- * @copyright	Copyright (c) 2008 - 2014, EllisLab, Inc. (http://ellislab.com/)
- * @copyright	Copyright (c) 2014 - 2015, British Columbia Institute of Technology (http://bcit.ca/)
- * @license	http://opensource.org/licenses/MIT	MIT License
- * @link	http://codeigniter.com
- * @since	Version 1.0.0
- * @filesource
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
  */
-defined('BASEPATH') OR exit('No direct script access allowed');
+
+use Config\Services;
 
 /**
  * CodeIgniter Security Helpers
- *
- * @package		CodeIgniter
- * @subpackage	Helpers
- * @category	Helpers
- * @author		EllisLab Dev Team
- * @link		http://codeigniter.com/user_guide/helpers/security_helper.html
  */
 
-// ------------------------------------------------------------------------
-
-if ( ! function_exists('xss_clean'))
+if (! function_exists('sanitize_filename'))
 {
 	/**
-	 * XSS Filtering
+	 * Sanitize a filename to use in a URI.
 	 *
-	 * @param	string
-	 * @param	bool	whether or not the content is an image file
-	 * @return	string
+	 * @param string $filename
+	 *
+	 * @return string
 	 */
-	function xss_clean($str, $is_image = FALSE)
+	function sanitize_filename(string $filename): string
 	{
-		return get_instance()->security->xss_clean($str, $is_image);
+		return Services::security()->sanitizeFilename($filename);
 	}
 }
 
-// ------------------------------------------------------------------------
+//--------------------------------------------------------------------
 
-if ( ! function_exists('sanitize_filename'))
-{
-	/**
-	 * Sanitize Filename
-	 *
-	 * @param	string
-	 * @return	string
-	 */
-	function sanitize_filename($filename)
-	{
-		return get_instance()->security->sanitize_filename($filename);
-	}
-}
-
-// --------------------------------------------------------------------
-
-if ( ! function_exists('do_hash'))
-{
-	/**
-	 * Hash encode a string
-	 *
-	 * @todo	Remove in version 3.1+.
-	 * @deprecated	3.0.0	Use PHP's native hash() instead.
-	 * @param	string	$str
-	 * @param	string	$type = 'sha1'
-	 * @return	string
-	 */
-	function do_hash($str, $type = 'sha1')
-	{
-		if ( ! in_array(strtolower($type), hash_algos()))
-		{
-			$type = 'md5';
-		}
-
-		return hash($type, $str);
-	}
-}
-
-// ------------------------------------------------------------------------
-
-if ( ! function_exists('strip_image_tags'))
+if (! function_exists('strip_image_tags'))
 {
 	/**
 	 * Strip Image Tags
 	 *
-	 * @param	string
-	 * @return	string
+	 * @param string $str
+	 *
+	 * @return string
 	 */
-	function strip_image_tags($str)
+	function strip_image_tags(string $str): string
 	{
-		return get_instance()->security->strip_image_tags($str);
+		return preg_replace([
+			'#<img[\s/]+.*?src\s*=\s*(["\'])([^\\1]+?)\\1.*?\>#i',
+			'#<img[\s/]+.*?src\s*=\s*?(([^\s"\'=<>`]+)).*?\>#i',
+		], '\\2', $str
+		);
 	}
 }
 
-// ------------------------------------------------------------------------
+//--------------------------------------------------------------------
 
-if ( ! function_exists('encode_php_tags'))
+if (! function_exists('encode_php_tags'))
 {
 	/**
 	 * Convert PHP tags to entities
 	 *
-	 * @param	string
-	 * @return	string
+	 * @param string $str
+	 *
+	 * @return string
 	 */
-	function encode_php_tags($str)
+	function encode_php_tags(string $str): string
 	{
-		return str_replace(array('<?', '?>'), array('&lt;?', '?&gt;'), $str);
+		return str_replace(['<?', '?>'], ['&lt;?', '?&gt;'], $str);
 	}
 }
+
+//--------------------------------------------------------------------
